refactored service return types

+++ account-service/src/main/java/com/rahim/accountservice/service/repository/implementation/AccountRepositoryHandlerService.java
@@ -8,17 +8,13 @@ import lombok.RequiredArgsConstructor;
 import org.hibernate.exception.DataException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
 
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
-import java.util.Collections;
 import java.util.List;
 
 /**
@@ -33,7 +29,6 @@ public class AccountRepositoryHandlerService implements IAccountRepositoryHandle
     private final AccountRepository accountRepository;
 
     @Override
-    @Transactional(readOnly = true)
     public Account findById(int accountId) {
         return accountRepository.findById(accountId).orElse(new Account());
     }
@@ -55,16 +50,15 @@ public class AccountRepositoryHandlerService implements IAccountRepositoryHandle
 
     @Override
     public void deleteAccount(int accountId) {
-        findById(accountId).ifPresent(userAccount -> {
-            try {
-                LOG.debug("Deleting account with ID: {}", accountId);
-                accountRepository.deleteById(accountId);
-                LOG.debug("Account with ID {} deleted successfully", accountId);
-            } catch (EmptyResultDataAccessException e) {
-                LOG.warn("Attempted to delete non-existing account with ID: {}", accountId);
-                throw new EntityNotFoundException("Account with ID " + accountId + " not found");
-            }
-        });
+        Account account = findById(accountId);
+
+        if (account.getId() == null) {
+            LOG.warn("Attempted to delete non-existing account with ID: {}", accountId);
+            throw new EntityNotFoundException("Account with ID " + accountId + " not found");
+        }
+
+        accountRepository.deleteById(accountId);
+        LOG.debug("Account with ID {} deleted successfully", accountId);
     }
 
     @Override
@@ -94,21 +88,7 @@ public class AccountRepositoryHandlerService implements IAccountRepositoryHandle
     @Override
     @Transactional(readOnly = true)
     public List<Account> getAllAccounts() {
-        List<Account> accounts = Collections.emptyList();
-
-        try {
-            accounts = accountRepository.findAll();
-            if (CollectionUtils.isEmpty(accounts)) {
-                LOG.debug("No accounts found in the database");
-            } else {
-                LOG.debug("Fetched {} accounts from the database", accounts.size());
-            }
-
-        } catch (DataAccessException e) {
-            LOG.error("Error occurred while fetching accounts from the database", e);
-        }
-
-        return accounts;
+        return accountRepository.findAll();
     }
 
     @Override
